refactor: 拆分公告界面组件到独立模块
