+++ src/renderer/components/states/announcement/AnnouncementContent.tsx
@@ -25,67 +25,28 @@
  */
 
 import type { ReactElement } from 'react';
-import { useTranslation } from 'react-i18next';
 import useIslandStore from '../../../store/slices';
+import { AnnouncementHeader } from './components/AnnouncementHeader';
+import { AnnouncementBody } from './components/AnnouncementBody';
 import { useAnnouncementData } from './hooks/useAnnouncementData';
 import '../../../styles/announcement/announcement.css';
-
-function formatDatetime(value?: string): string {
-  if (!value) return '';
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return value;
-  return `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, '0')}-${String(date.getDate()).padStart(2, '0')} ${String(date.getHours()).padStart(2, '0')}:${String(date.getMinutes()).padStart(2, '0')}`;
-}
 
 /**
  * 公告页内容组件
  * @returns 公告状态视图
  */
 export function AnnouncementContent(): ReactElement {
-  const { t } = useTranslation();
   const { setHover } = useIslandStore();
   const { loading, announcement } = useAnnouncementData();
 
   return (
     <div className="announcement-state-content" onClick={(event) => event.stopPropagation()}>
       <div className="announcement-panel">
-        <div className="announcement-panel-header">
-          <div>
-            <div className="announcement-title">{t('announcement.title', { defaultValue: '公告' })}</div>
-            <div className="announcement-subtitle">{t('announcement.subtitle', { defaultValue: '当前已是最新版本，以下为最新公告内容。' })}</div>
-          </div>
-          <div className="announcement-header-actions">
-            {announcement && (
-              <div className="announcement-meta-vertical">
-                <span className="announcement-meta-title">{announcement.title || t('announcement.defaultTitle', { defaultValue: '系统公告' })}</span>
-                {announcement.updatedAt && (
-                  <span className="announcement-meta-time">{t('announcement.updatedAt', { defaultValue: '更新时间：{{time}}', time: formatDatetime(announcement.updatedAt) })}</span>
-                )}
-              </div>
-            )}
-            <button type="button" className="announcement-close-btn" onClick={() => setHover()}>
-              {t('announcement.close', { defaultValue: '关闭' })}
-            </button>
-          </div>
-        </div>
+        <AnnouncementHeader announcement={announcement} onClose={() => setHover()} />
 
         <div className="announcement-divider" />
 
-        {loading ? (
-          <div className="announcement-empty">{t('announcement.loading', { defaultValue: '正在加载公告…' })}</div>
-        ) : !announcement ? (
-          <div className="announcement-empty">{t('announcement.empty', { defaultValue: '暂无公告内容' })}</div>
-        ) : (
-          <>
-            {announcement.contentHtml ? (
-              <div className="announcement-body" dangerouslySetInnerHTML={{ __html: announcement.contentHtml }} />
-            ) : (
-              <div className="announcement-body">
-                <pre>{announcement.content || ''}</pre>
-              </div>
-            )}
-          </>
-        )}
+        <AnnouncementBody loading={loading} announcement={announcement} />
       </div>
     </div>
   );
